BUG: finally got it to load eggnog.tests(you need an __init__
