+++ q2_types_genomics/eggnog/tests/test_formats.py
@@ -14,7 +14,10 @@ from q2_types_genomics.eggnog import (
 
 from qiime2.plugin.testing import TestPluginBase
 
+import os
 import shutil
+
+# Small, example databases were downloaded from http://eggnog5.embl.de/#/app/downloads
 
 
 class TestRefFmts(TestPluginBase):
@@ -29,15 +32,18 @@ class TestRefFmts(TestPluginBase):
 
     bad_file_names = [
         "eggnog.db", "eggnog.taxa.db", "eggnog.taxa.db.traverse.pkl",
-        "eggnog_proteins.dmnd", "novel_fams.dmnd", "diamond_test.dmnd"
+        "eggnog_proteins.dmnd", "novel_fams.dmnd", "diamond_testing.dmnd"
         ]
 
+    # this will fail right now because Directory formats expect all files in
+    # the format definition to be present. Going to work on making optional.
     def test_registration_basics(self):
-        filename = 'diamond_test.dmnd'
+        filename = 'diamond_testing.dmnd'
         filepath = self.get_data_path(filename)
 
-        fmt = EggnogRefDirFmt(filepath, 'r')
-        fmt.validate(level='max')
+        fmt = EggnogRefDirFmt(mode='w')
+        shutil.copy(src=filepath, dst=fmt.path)
+        fmt.validate()
 
 #    def test_fails_on_string(self):
 #        filename = 'sample_annotations.txt'
@@ -53,3 +59,27 @@ def _artifact_generator(*data):
     for filename in filenames:
         fp = self.get_data_path(filename)
         print(fp)
+
+
+class TestEggnogOutput(TestPluginBase):
+    package='q2_types_genomics.eggnog.tests'
+
+    def test_initial_setup(self):
+        # Create directory
+        test_output = EggnogOutputDirFmt()
+
+        # trying putting a file in...
+        shutil.copy(self.get_data_path("sample_eggnog_annotations.annotations"),
+                    test_output.path)
+
+        shutil.copy(self.get_data_path("sample_annotations.txt"),
+                    test_output.path)
+        test_output.save(self.get_data_path("test_output.qza"))
+        artifacted = _return_artifact(test_output)
+
+        fh = Artifact.load(artifacted)
+        self.assertIn("sample_annotations.txt", fh)
+
+
+def _return_artifact(fmtd_object):
+    return fmtd_object
