+++ setup.py
@@ -38,7 +38,9 @@ setup(
             ['data/*', 'data/genes/*', 'data/loci/*',
              'data/loci-invalid/*', 'data/proteins/*'],
         'q2_types_genomics.eggnog.tests':
-            ['data/*', ],
+            ['data/*'],
+        #'q2_types_genomics.eggnog.tests':
+        #    ['data/*', ],
 
     },
     zip_safe=False,
